requirements.txt
asgiref==3.8.1
colorama==0.4.6
Django==5.1.6
django-cors-headers==4.7.0
django-debug-toolbar==5.2.0
django-easy-audit==1.3.7
django-filter==25.1
djangorestframework==3.15.2
djangorestframework_simplejwt==5.5.0
drf-yasg==1.21.9
Faker==36.1.1
importlib_metadata==8.6.1
inflection==0.5.1
Markdown==3.7
packaging==24.2
pillow==11.1.0
psycopg2-binary==2.9.10
PyJWT==2.9.0
python-decouple==3.8
pytz==2025.1
PyYAML==6.0.2
qrcode==8.0
rest-framework-simplejwt==0.0.2
sqlparse==0.5.3
typing_extensions==4.12.2
tzdata==2025.1
uritemplate==4.1.1
zipp==3.21.0
